requirements.txt
fastapi
uvicorn
pandas
google-cloud-storage
gcsfs
selenium
chromedriver-binary
webdriver-manager
beautifulsoup4